requirements.txt
livekit-agents[openai,silero,deepgram]~=1.0
python-dotenv
black
pyyaml